+++ src/industry-solution-packs-composable-applications/index.js
@@ -1,3 +1,6 @@
 export * from './industry-solution-packs-composable-applications-overview-constants.js';
 export { IndustrySolutionPacksComposableApplicationsOverviewProfile } from './contracts/industry-solution-packs-composable-applications-overview-profile.js';
 export { IndustrySolutionPacksComposableApplicationsOverviewDescriptor } from './overview/industry-solution-packs-composable-applications-overview-descriptor.js';
+export * from './solution-pack-model-boundaries-manifests-constants.js';
+export { SolutionPackModelBoundariesManifestsProfile } from './contracts/solution-pack-model-boundaries-manifests-profile.js';
+export { SolutionPackModelBoundariesManifestsDescriptor } from './manifests/solution-pack-model-boundaries-manifests-descriptor.js';